Move and/or-related dbAdapter helpers to the utils file
