+++ app/support/DbAdapter/utils.js
@@ -98,3 +98,29 @@ export function withDbHelpers(db) {
   Object.setPrototypeOf(wrapper, db); // eslint-disable-line prefer-reflect
   return wrapper;
 }
+
+function joinThem(array, joinBy, defaultValue, shortcutValue, skipValue) {
+  if (array.some((x) => x === shortcutValue)) {
+    return shortcutValue;
+  }
+
+  const parts = array.filter((x) => !!x && x !== skipValue);
+
+  if (parts.length === 0) {
+    return defaultValue;
+  }
+
+  if (parts.length === 1) {
+    return parts[0];
+  }
+
+  return `(${parts.join(` ${joinBy} `)})`;
+}
+
+export function andJoin(array, def = 'true') {
+  return joinThem(array, 'and', def, 'false', 'true');
+}
+
+export function orJoin(array, def = 'false') {
+  return joinThem(array, 'or', def, 'true', 'false');
+}
